Persist context for pushing notifications
